avatar/interactive/api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azurefunctions-extensions-http-fastapi
openai
pyodbc
requests
beautifulsoup4
azure-ai-projects~=1.0.0b7
azure-identity